am_framework/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
pycparser==2.22
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
mysqlclient==2.2.4
coverage==7.4.3
black==24.2.0
flake8==7.0.0
gunicorn==22.0.0
mysql-connector-python==8.3.0
mypy==1.10.0